extracted TimeConfig enum
