+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/gemini/editor/targetComponent/EdCompTargetList.java
@@ -96,7 +96,7 @@ public final class EdCompTargetList extends OtItemEditor<ISPObsComponent, Target
     private final TelescopeForm            _w;
     private final NonSiderealTargetSupport _nonSiderealTargetSup;
     private final TimeDocument             _timeDocument;
-
+//
     // Stuff that varies with time
     private SPTarget        _curPos;
     private GuideGroup      _curGroup;
@@ -1340,7 +1340,6 @@ public final class EdCompTargetList extends OtItemEditor<ISPObsComponent, Target
         // Update target details
         _detailEditor.edit(getObsContext(env), _curPos);
 
-
     }
 
 
@@ -1975,59 +1974,3 @@ enum UserPositionType implements PositionType {
 }
 
 
-/**
- * The time configurations are pre-sets of dates the Horizons query should use to gets its
- * information.
- */
-enum TimeConfig {
-
-    NOW("Now") {
-        public Date getDate() {
-            return new Date();
-        }
-    },
-    ONE_HOUR("1 Hour") {
-        public Date getDate() {
-            final Date d = new Date();
-            return new Date(d.getTime() + HOUR);
-        }
-    },
-    TWO_HOUR("2 Hours") {
-        public Date getDate() {
-            final Date d = new Date();
-            return new Date(d.getTime() + HOUR * 2);
-        }
-    },
-
-    THREE_HOUR("3 Hours") {
-        public Date getDate() {
-            final Date d = new Date();
-            return new Date(d.getTime() + HOUR * 3);
-        }
-    },
-
-    FIVE_HOUR("5 Hours") {
-        public Date getDate() {
-            final Date d = new Date();
-            return new Date(d.getTime() + HOUR * 5);
-        }
-    },;
-
-    private static final int HOUR = 1000 * 60 * 60;
-    private final String _displayValue;
-
-    private TimeConfig(String displayValue) {
-        _displayValue = displayValue;
-    }
-
-    public String displayValue() {
-        return _displayValue;
-    }
-
-    /**
-     * Return the <code>Date</code>  for the given configuration
-     * @return the <code>Date</code> associated to this configuration
-     */
-    public abstract Date getDate();
-
-}
